Abstract and improve HousingCompany serializers
- Also simplify other serializers by using the correct serializers for their fields
- Move Building and RealEstate serializers to their own files
