+++ backend/hitas/views/codes.py
@@ -1,20 +1,12 @@
-from typing import Optional
-
 from rest_framework import serializers
 
 from hitas.models.codes import AbstractCode, BuildingType, Developer, FinancingMethod
-from hitas.views.helpers import value_or_none
+from hitas.views.helpers import ValueOrNullField
 
 
 class AbstractCodeSerializer(serializers.ModelSerializer):
-    description = serializers.SerializerMethodField()
+    description = ValueOrNullField()
     code = serializers.CharField(source="legacy_code_number")
-
-    def get_description(self, obj: AbstractCode) -> Optional[str]:
-        return value_or_none(obj.description)
-
-    def get_code(self, obj: AbstractCode) -> Optional[str]:
-        return obj.legacy_code_number
 
     class Meta:
         model = AbstractCode
